CV-Reformatter-POC/requirements.txt
python-docx>=1.2.0
mammoth>=1.11.0
pytest>=8.0.0
